+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Default` for `GroupSorted`
 - `Default` for `ProductSorted`
 - `Default` for `ZipSorted`
+- `Default` for `Compress`
 
 ### Changed